+++ src/index.js
@@ -7,10 +7,8 @@ import { baseStyles, fnAnsi256, fnBgAnsi256, fnRgb, fnBgRgb } from './ansi-codes
  * @property {string} close
  * @property {string?} openStack
  * @property {string?} closeStack
- * @property {null | AnsisProps} parent
+ * @property {null | AnsisProps} props
  */
-
-//Object.defineProperty(exports, '__esModule', { value: !0 });
 
 const { defineProperty, defineProperties, setPrototypeOf } = Object;
 
@@ -84,7 +82,7 @@ const createStyle = ({ props }, { open, close }) => {
   }
 
   setPrototypeOf(style, stylePrototype);
-  style.props = { open, close, openStack, closeStack, parent: props };
+  style.props = { open, close, openStack, closeStack, props: props };
   style.open = openStack;
   style.close = closeStack;
 
@@ -108,7 +106,7 @@ const wrap = (strings, values, props) => {
   if (~string.indexOf('\x1b')) {
     while (props !== undefined) {
       string = strReplaceAll(string, props.close, props.open);
-      props = props.parent;
+      props = props.props;
     }
   }
 
@@ -156,70 +154,73 @@ const ansis = new Ansis();
 
 export { Ansis, ansis as default };
 
-export const ansi256 = ansis.ansi256;
-export const ansi = ansis.ansi;
-export const fg = ansis.fg;
-export const bgAnsi256 = ansis.bgAnsi256;
-export const bgAnsi = ansis.bgAnsi;
-export const bg = ansis.bg;
-export const rgb = ansis.rgb;
-export const bgRgb = ansis.bgRgb;
-export const hex = ansis.hex;
-export const bgHex = ansis.bgHex;
+export const {
+  // color functions
+  ansi256,
+  ansi,
+  fg,
+  bgAnsi256,
+  bgAnsi,
+  bg,
+  rgb,
+  bgRgb,
+  hex,
+  bgHex,
 
-// misc
-export const reset = ansis.reset;
-export const inverse = ansis.inverse;
-export const hidden = ansis.hidden;
-export const visible = ansis.visible;
+  // misc
+  reset,
+  inverse,
+  hidden,
+  visible,
 
-// styles
-export const bold = ansis.bold;
-export const dim = ansis.dim;
-export const faint = ansis.faint;
-export const italic = ansis.italic;
-export const underline = ansis.underline;
-export const doubleUnderline = ansis.doubleUnderline;
-export const strikethrough = ansis.strikethrough;
-export const strike = ansis.strike;
-export const frame = ansis.frame;
-export const encircle = ansis.encircle;
-export const overline = ansis.overline;
+  // styles
+  bold,
+  dim,
+  faint,
+  italic,
+  underline,
+  doubleUnderline,
+  strikethrough,
+  strike,
+  frame,
+  encircle,
+  overline,
 
-// foreground colors
-export const black = ansis.black;
-export const red = ansis.red;
-export const green = ansis.green;
-export const yellow = ansis.yellow;
-export const blue = ansis.blue;
-export const magenta = ansis.magenta;
-export const cyan = ansis.cyan;
-export const white = ansis.white;
-export const gray = ansis.gray;
-export const grey = ansis.grey;
-export const blackBright = ansis.blackBright;
-export const redBright = ansis.redBright;
-export const greenBright = ansis.greenBright;
-export const yellowBright = ansis.yellowBright;
-export const blueBright = ansis.blueBright;
-export const magentaBright = ansis.magentaBright;
-export const cyanBright = ansis.cyanBright;
-export const whiteBright = ansis.whiteBright;
+  // foreground colors
+  black,
+  red,
+  green,
+  yellow,
+  blue,
+  magenta,
+  cyan,
+  white,
+  gray,
+  grey,
+  blackBright,
+  redBright,
+  greenBright,
+  yellowBright,
+  blueBright,
+  magentaBright,
+  cyanBright,
+  whiteBright,
 
-// background colors
-export const bgBlack = ansis.bgBlack;
-export const bgRed = ansis.bgRed;
-export const bgGreen = ansis.bgGreen;
-export const bgYellow = ansis.bgYellow;
-export const bgBlue = ansis.bgBlue;
-export const bgMagenta = ansis.bgMagenta;
-export const bgCyan = ansis.bgCyan;
-export const bgWhite = ansis.bgWhite;
-export const bgBlackBright = ansis.bgBlackBright;
-export const bgRedBright = ansis.bgRedBright;
-export const bgGreenBright = ansis.bgGreenBright;
-export const bgYellowBright = ansis.bgYellowBright;
-export const bgBlueBright = ansis.bgBlueBright;
-export const bgMagentaBright = ansis.bgMagentaBright;
-export const bgCyanBright = ansis.bgCyanBright;
-export const bgWhiteBright = ansis.bgWhiteBright;
+  // background colors
+  bgBlack,
+  bgRed,
+  bgGreen,
+  bgYellow,
+  bgBlue,
+  bgMagenta,
+  bgCyan,
+  bgWhite,
+  bgBlackBright,
+  bgRedBright,
+  bgGreenBright,
+  bgYellowBright,
+  bgBlueBright,
+  bgMagentaBright,
+  bgCyanBright,
+  bgWhiteBright,
+} = ansis;